Account for empty variant arrays in parser
It's possible for entries in `.shadervariants` assets to have an empty variants array. Eg. `variants: []`.

This causes the parser to go out-of-bounds in the `yaml` list when the final entry has an empty variants array.
When not the final entry, it probably causes the next entry to be parsed incorrectly.